changed surface cost from double to int

+++ ddl.sql
@@ -116,7 +116,7 @@ CREATE TABLE surface (
   pool_id BIGINT REFERENCES pool(id),
   installed INT NOT NULL,
   kind VARCHAR NOT NULL,
-  cost NUMERIC(7, 2) NOT NULL
+  cost INT NOT NULL
 );
 
 CREATE TABLE deck (
